+++ listings/ch09/listing_14.js
@@ -1,3 +1,22 @@
-export function toPromise(value) {
-  return Promise.resolve(value)
+export function mountDOM(vdom, parentEl, index/*--add--*/, hostComponent = null/*--add--*/) {
+  switch (vdom.type) {
+    case DOM_TYPES.TEXT: {
+      createTextNode(vdom, parentEl, index)
+      break
+    }
+
+    case DOM_TYPES.ELEMENT: {
+      createElementNode(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
+      break
+    }
+
+    case DOM_TYPES.FRAGMENT: {
+      createFragmentNodes(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
+      break
+    }
+
+    default: {
+      throw new Error(`Can't mount DOM of type: ${vdom.type}`)
+    }
+  }
 }
