+++ listings/ch09/listing_16.js
@@ -1,10 +1,6 @@
-function createElementNode(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/) {
-  const { tag, props, children } = vdom
+function addProps(el, props, vdom/*--add--*/, hostComponent/*--add--*/) {
+  const { on: events, ...attrs } = props
 
-  const element = document.createElement(tag)
-  addProps(element, props, vdom/*--add--*/, hostComponent/*--add--*/)
-  vdom.el = element
-
-  children.forEach((child) => mountDOM(child, element/*--add--*/, null, hostComponent/*--add--*/))
-  insert(element, parentEl, index)
+  vdom.listeners = addEventListeners(events, el/*--add--*/, hostComponent/*--add--*/)
+  setAttributes(el, attrs)
 }
