+++ listings/ch09/listing_18.js
@@ -1,8 +1,27 @@
-function createFragmentNodes(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/) {
-  const { children } = vdom
-  vdom.el = parentEl
+export function patchDOM(oldVdom, newVdom, parentEl, hostComponent = null) {
+  if (!areNodesEqual(oldVdom, newVdom)) {
+    const index = Array.from(parentEl.childNodes).indexOf(oldVdom.el)
+    destroyDOM(oldVdom)
+    mountDOM(newVdom, parentEl, index/*--add--*/, hostComponent/*--add--*/)
 
-  children.forEach((child) =>
-    mountDOM(child, parentEl, index ? index + i : null/*--add--*/, hostComponent/*--add--*/)
-  )
+    return newVdom
+  }
+
+  newVdom.el = oldVdom.el
+
+  switch (newVdom.type) {
+    case DOM_TYPES.TEXT: {
+      patchText(oldVdom, newVdom)
+      return newVdom
+    }
+
+    case DOM_TYPES.ELEMENT: {
+      patchElement(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/)
+      break
+    }
+  }
+
+  patchChildren(oldVdom, newVdom, hostComponent)
+
+  return newVdom
 }
