+++ listings/ch09/listing_20.js
@@ -1,29 +1,30 @@
-function patchElement(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/) {
-  const el = oldVdom.el
-  const {
-    class: oldClass,
-    style: oldStyle,
-    on: oldEvents,
-    ...oldAttrs
-  } = oldVdom.props
-  const {
-    class: newClass,
-    style: newStyle,
-    on: newEvents,
-    ...newAttrs
-  } = newVdom.props
-  const { listeners: oldListeners } = oldVdom
+function patchEvents(
+  el,
+  oldListeners = {},
+  oldEvents = {},
+  newEvents = {},
+  // --add--
+  hostComponent
+  // --add--
+) {
+  const { removed, added, updated } = objectsDiff(oldEvents, newEvents)
 
-  patchAttrs(el, oldAttrs, newAttrs)
-  patchClasses(el, oldClass, newClass)
-  patchStyles(el, oldStyle, newStyle)
-  newVdom.listeners = patchEvents(
-    el,
-    oldListeners,
-    oldEvents,
-    newEvents,
-    // --add--
-    hostComponent
-    // --add--
-  )
+  for (const eventName of removed.concat(updated)) {
+    el.removeEventListener(eventName, oldListeners[eventName])
+  }
+  const addedListeners = {}
+
+  for (const eventName of added.concat(updated)) {
+    const listener = addEventListener(
+      eventName,
+      newEvents[eventName],
+      el,
+      // --add--
+      hostComponent
+      // --add--
+    )
+    addedListeners[eventName] = listener
+  }
+
+  return addedListeners
 }
